Move CreateVIPNetworkList to common_openstackcontrolplane.go
It depends on the controlplane as input, so is dependet on the CRD
version.
